// code-app/src/generated/models/Sams_usecasesystemsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Sams_usecasesystemssams_access = {
  726410000: 'Read',
  726410001: 'Write',
  726410002: 'Readandwrite',
  726410003: 'Notsure',
} as const;
export type Sams_usecasesystemssams_access = keyof typeof Sams_usecasesystemssams_access;
export const Sams_usecasesystemssams_connectoravailable = {
  726410000: 'Yes',
  726410001: 'No',
  726410002: 'Notsure',
} as const;
export type Sams_usecasesystemssams_connectoravailable =
  keyof typeof Sams_usecasesystemssams_connectoravailable;
export const Sams_usecasesystemssams_systemtype = {
  726410000: 'Microsoft365app',
  726410001: 'DataverseorDynamics365',
  726410002: 'SAP',
  726410003: 'ServiceNow',
  726410004: 'Salesforce',
  726410005: 'CustomAPIordatabase',
  726410006: 'Websiteorportal',
  726410007: 'Other',
} as const;
export type Sams_usecasesystemssams_systemtype = keyof typeof Sams_usecasesystemssams_systemtype;
export const Sams_usecasesystemsstatuscode = {
  1: 'Active',
  2: 'Inactive',
} as const;
export type Sams_usecasesystemsstatuscode = keyof typeof Sams_usecasesystemsstatuscode;

export interface Sams_usecasesystemsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  sams_access?: Sams_usecasesystemssams_access;
  sams_connectoravailable?: Sams_usecasesystemssams_connectoravailable;
  sams_name: string;
  sams_systemtype?: Sams_usecasesystemssams_systemtype;
  'sams_UseCaseId@odata.bind': string;
  sams_usecasesystemid: string;
  statecode: string;
  statuscode?: Sams_usecasesystemsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Sams_usecasesystems extends Sams_usecasesystemsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  sams_accessname?: string;
  sams_connectoravailablename?: string;
  sams_systemtypename?: string;
  sams_usecaseidname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  sams_usecaseid?: object;
  _sams_usecaseid_value?: string;
}